Amend feedback from PR
